fix(cudf): correct hash_with_seed return type from bigint to integer
cudf::hashing::murmurhash3_x86_32 returns uint32 (INT32), matching
Spark's expected INTEGER return type for hash functions. The previous
bigint signature caused signature mismatch, making the function
unrecognizable and falling back to CPU for Q8, Q14, Q38, Q47, Q57, Q87.

Made-with: Cursor

+++ velox/experimental/cudf/expression/ExpressionEvaluator.cpp
@@ -1054,7 +1054,7 @@ bool registerBuiltinFunctions(const std::string& prefix) {
         return std::make_shared<HashFunction>(expr);
       },
       {FunctionSignatureBuilder()
-           .returnType("bigint")
+           .returnType("integer")
            .constantArgumentType("integer")
            .argumentType("any")
            .variableArity()
